requirements.txt
numpy==1.19.5
pandas==1.1.5
pymongo==4.0.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
